Include "choose" in empty_vales parameter (refactor)
Added the string "choose" to be included in the default value for empty_values in the QueryParser._parse_where_part function. When trying to query the basic query form, the value "choose" would be the default value rather than "choose a value" and this was a quick fix at handling the issue.