Initial Commit (feature)
SQL used to create a consituency_list helper view and an advance_data view which is used to in the query of the database. Changed some joins to be inner joins, added consituency_list_view to pull states and districts together in relation to person id.

Added columns party_id,state_id,district_id to be used in the query rather than
their string counterparts.