Add functionality to "append" helper tables to query (feature)
A HelperTableMap dataclass was created to house instructions to make changes to the query if provided to the function QueryFilter.remap_query_inputs. If provided a the table name will append itself to the column to be queried upon or a provided column through setting the select_to_change attribute. Joins were created in the QueryParse._create_joins by using the column_to_join attribute in the dataclass. This essentially allows other "helper" tables/views to be joined to the query and assign to the appropriate column.